security_context_v1: set CLOEXEC for client FDs
Closes: https://gitlab.freedesktop.org/wlroots/wlroots/-/work_items/3949
(cherry picked from commit 327f000532e96329280e06eb12af2d98071ac023)
