+++ 3rdparty/wlroots/include/util/fd.h
@@ -1,0 +1,8 @@
+#ifndef UTIL_FD_H
+#define UTIL_FD_H
+
+#include <stdbool.h>
+
+bool set_cloexec(int fd, bool cloexec);
+
+#endif
