+++ 3rdparty/wlroots/xwayland/sockets.h
@@ -3,7 +3,6 @@
 
 #include <stdbool.h>
 
-bool set_cloexec(int fd, bool cloexec);
 void unlink_display_sockets(int display);
 int open_display_sockets(int socks[2]);
